% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairsD3.R
\name{pairsD3Output}
\alias{pairsD3Output}
\title{Widget output function for use in Shiny}
\usage{
pairsD3Output(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Shiny output ID}

\item{width}{width default '100\%'}

\item{height}{height default '400px'}
}
\description{
Widget output function for use in Shiny
}

