% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinypairs.R
\name{shinypairs}
\alias{shinypairs}
\title{Shiny interface to the pairsD3 function}
\usage{
shinypairs(x, group = NULL, subset = NULL, labels = NULL)
}
\arguments{
\item{x}{the coordinates of points given as numeric columns of a
matrix or data frame. Logical and factor columns are converted
to numeric in the same way that \code{data.matrix} does.}

\item{group}{a optional vector specifying the group each observation
belongs to.  Used for tooltips and colouring the observations.}

\item{subset}{an optional vector specifying a subset of observations
to be used for plotting. Useful when you have a large number of
observations, you can specify a random subset.}

\item{labels}{the names of the variables (column names of \code{x}
used by default).}
}
\description{
Opens a shiny GUI to facilitate interaction with the pairsD3 function
}
\examples{
data(iris)
\dontrun{
shinypairs(iris)
}

}

