% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pairsD3.R
\name{renderPairsD3}
\alias{renderPairsD3}
\title{Widget render function for use in Shiny}
\usage{
renderPairsD3(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

